Cleanup and fix bugs
